Large variables refactor (untested)

+++ monroe-experiments/usr/bin/container-stop.sh
@@ -3,18 +3,34 @@
 
 SCHEDID=$1
 STATUS=$2
-CONTAINER=monroe-$SCHEDID
 
-BASEDIR=/experiments/user
-STATUSDIR=$BASEDIR
+# Default variables
+USERDIR="/experiments/user"
+CONTAINER_NAME=monroe-$SCHEDID
 USAGEDIR=/monroe/usage/netns
-
 MNS="ip netns exec monroe"
+
+NEAT_CONTAINER_NAME=monroe-neat-proxy
+EXCLUDED_IF="Br|lo|metadata|wwan|ifb|docker"
+OPINTERFACES="nlw_|wwan"
+URL_NEAT_PROXY=monroe/neat-proxy
+
+# Update above default variables if needed 
+. /etc/default/monroe-experiments
+
+_tmpfile_=/tmp/cleanup-$SCHEDID.log
+
+exec &> >(tee -a $_tmpfile_) || {
+   echo "Could not create log file $_tmpfile_"
+}
+
 VTAPPREFIX=mvtap-
 VM_TMP_FILE=/experiments/virtualization.rd/$SCHEDID.tar_dump
+VM_OS_DISK=/var/lib/docker/scratch/virtualization/image-$SCHEDID.qcow2
 
-if [ -f $BASEDIR/$SCHEDID.conf ]; then
-  CONFIG=$(cat $BASEDIR/$SCHEDID.conf);
+
+if [ -f $USERDIR/$SCHEDID.conf ]; then
+  CONFIG=$(cat $USERDIR/$SCHEDID.conf);
   IS_INTERNAL=$(echo $CONFIG | jq -r '.internal // empty');
   STARTTIME=$(echo $CONFIG | jq -r '.start // empty');
   BDEXT=$(echo $CONFIG | jq -r '.basedir // empty');
@@ -22,39 +38,33 @@ if [ -f $BASEDIR/$SCHEDID.conf ]; then
   IS_VM=$(echo $CONFIG | jq -r '.vm // empty');
   NEAT_PROXY=$(echo $CONFIG | jq -r '.neat // empty');
 fi
+#TODO: STOP an or here?
+
+
 if [ ! -z "$IS_INTERNAL" ]; then
-  BASEDIR=/experiments/monroe${BDEXT}
-fi
-
-VM_OS_DISK=/var/lib/docker/scratch/virtualization/image-$SCHEDID.qcow2
-
-exec > /tmp/cleanup.log 2>&1
-
-CID=$( docker ps -a | grep $CONTAINER | awk '{print $1}' )
-
-echo "Finalize accounting."
-/usr/bin/usage-defaults
-
-
-echo -n "Stopping container... "
-if [ $(docker inspect -f "{{.State.Running}}" $CID 2>/dev/null) ]; then
-  RUNNING=$(docker inspect $CID|jq -r .[].State.Status)
-  if [ "$RUNNING" == "exited" ]; then
-    if [ -z "$STATUS" ]; then
-      STATUS="finished"
-    fi
-  fi
-
-  docker stop --time=10 $CID;
-  echo "stopped:"
-  docker inspect $CID|jq .[].State
+  _EXPPATH=/experiments/monroe${BDEXT}/$SCHEDID
 else
-  echo "Container is no longer running.";
+  _EXPPATH=$USERDIR/$SCHEDID
 fi
 
-if [[ -f $BASEDIR/$SCHEDID.pid && -z "$RUNNING" ]]; then 
+[ -x /usr/bin/usage-defaults ] && {
+  echo "Finalize accounting."
+  /usr/bin/usage-defaults
+}
+
+if [ docker inspect $CONTAINER_NAME &>/dev/null ]; then 
+  echo -n "Stopping container... "
+  if [ "$(docker inspect -f "{{.State.Running}}" $CONTAINER_NAME) &>/dev/null" == "true" ]; then
+    docker stop --time=10 $CONTAINER_NAME
+  fi
+  echo "stopped:"
+  docker inspect $CONTAINER_NAME|jq .[].State
+  if [ -z "$STATUS" ]; then
+    STATUS="finished"
+  fi
+elif [[ -f $_EXPPATH.pid ]]; then 
   echo -n "Killing vm (if any)... "
-  PID=$(cat $BASEDIR/$SCHEDID.pid)
+  PID=$(cat $_EXPPATH.pid)
   kill -9 $PID  # Should be more graceful maybe
   echo "ok."
 fi
@@ -72,7 +82,7 @@ if [[ -f $VM_TMP_FILE ]]; then # This file should NOT be here normaly
   echo "ok."
 fi
 
-VMIFHASH="$(cat $BASEDIR/$SCHEDID.vmifhash) 2>/dev/null"
+VMIFHASH="$(cat $_EXPPATH.vmifhash) 2>/dev/null"
 if [[ ! -z "$VMIFHASH" ]]; then
   VTAPS=$($MNS ls /sys/class/net/|grep "${VTAPPREFIX}${VMIFPREFIX}-")
   if [[ ! -z "$VTAPS" ]]; then 
@@ -94,73 +104,84 @@ if [ -f "/usr/bin/ykushcmd" ];then
     done
   fi
 fi
+
 ## Disable NEAT proxy ###
-if [[ ! -z "$NEAT_PROXY" ]]; then # If this is a experiment using the neat-proxy
+if [ ! -z "$NEAT_PROXY"  ] && [ -x /usr/bin/monroe-neat-init ]; then # If this is a experiment using the neat-proxy
   rm -f /etc/circle.d/60-*-neat-proxy.rules
   circle start
-  ## Stop the neta proxy container 
-  docker stop --time=10 monroe-neat-proxy
+  ## Stop the neat proxy container 
+  # TODO : remove container ?
+  docker stop --time=10 $NEAT_CONTAINER_NAME
 fi
 
-sysevent -t Scheduling.Task.Stopped -k id -v $SCHEDID
+[ -x /usr/bin/sysevent ] && sysevent -t Scheduling.Task.Stopped -k id -v $SCHEDID
 
-if [ -d $BASEDIR/$SCHEDID ]; then
+## Retrive Logs 
+if [ -d $_EXPPATH ]; then
   if [ -z "$IS_VM" ]; then 
     echo "Retrieving container logs:"
-    if [ ! -z "$CID" ]; then
-      docker logs -t $CID &> $STATUSDIR/$SCHEDID/container.log;
-    else
-      echo "CID not found for $CONTAINER." > $STATUSDIR/$SCHEDID/container.log;
-    fi
+    docker logs -t $CONTAINER_NAME &> $_EXPPATH/container.log;
     echo ""
   fi
   if [ ! -z "$STARTTIME" ]; then
+    # TODO : USe docker inspect to get time when container was running
     echo "Retrieving dmesg events:"
     dmesg|awk '{time=0 + substr($1,2,length($1)-2); if (time > '$STARTTIME') print $0}'
     echo ""
   fi
 
   echo -n "Syncing traffic statistics... "
-  monroe-user-experiments;
-  TRAFFIC=$(cat $STATUSDIR/$SCHEDID.traffic)
+  monroe-user-experiments
+  TRAFFIC=$(cat $_EXPPATH.traffic)
 
-  for i in $(ls $USAGEDIR/$SCHEDID/*.rx.total|sort); do
-    MACICCID=$(basename $i | sed -e 's/\..*//g')
-    TRAFFIC=$(echo "$TRAFFIC" | jq ".interfaces.\"$MACICCID\"=$(cat $USAGEDIR/$SCHEDID/$MACICCID.total)")
-  done;
+  # Will only exist if /usr/bin/usage has been running
+  if [ -d $USAGEDIR/$SCHEDID ];then 
+    for i in $(ls $USAGEDIR/$SCHEDID/*.rx.total|sort); do
+      MACICCID=$(basename $i | sed -e 's/\..*//g')
+      TRAFFIC=$(echo "$TRAFFIC" | jq ".interfaces.\"$MACICCID\"=$(cat $USAGEDIR/$SCHEDID/$MACICCID.total)")
+    done
+  fi
   if [ ! -z "$TRAFFIC" ]; then
-    echo "$TRAFFIC" > $STATUSDIR/$SCHEDID.traffic
-    echo "$TRAFFIC" > $STATUSDIR/$SCHEDID/container.stat
+    echo "$TRAFFIC" > $_EXPPATH.traffic
+    echo "$TRAFFIC" > $_EXPPATH/container.stat
   fi
   echo "ok."
 fi
 
+#TODO: What is this used for? 
 if [ -z "$STATUS" ]; then
-  echo 'stopped' > $STATUSDIR/$SCHEDID.status;
+  echo 'stopped' > $_EXPPATH.status;
 else
-  echo $STATUS > $STATUSDIR/$SCHEDID.status;
+  echo $STATUS > $_EXPPATH.status;
 fi
 
-echo -n "Untagging container image... "
-REF=$( docker images | grep $CONTAINER | awk '{print $3}' )
+echo -n "Deleting container image... "
+REF=$( docker images | grep $CONTAINER_NAME | awk '{print $3}' )
 if [ -z "$REF" ]; then
   echo "Container is no longer deployed.";
 else
-  docker rmi -f $CONTAINER
+  docker rmi -f $CONTAINER_NAME
 fi
 echo "ok."
 
+#TODO_ SHould this be here or in pull-base-containers et al.
 echo -n "Cleaning unused container images... "
 # remove all stopped containers (remove all, ignore errors when running)
+#TODO : Remove only monroe containers
 docker rm $(docker ps -aq) 2>/dev/null
 # clean any untagged containers without dependencies (unused layers)
 docker rmi $(docker images -a|grep '^<none>'|awk "{print \$3}") 2>/dev/null
 echo "ok."
 
-if [ ! -z "$EDUROAM_IDENTITY" ]; then
+#TODO: Make this dynamic to not hardcode to wlan0  
+EDUROAM_PID="$(pgrep -f /sbin/wpa_supplicant 2>/dev/null)"
+if [ ! -z "$EDUROAM_IDENTITY" ] && \
+   [ -x /usr/bin/eduroam-login.sh ] && \
+   [ ! -z "$EDUROAM_PID" ] && \
+   [ -z "$(ip link |grep wlan0:)"];then
     echo -n "Deleting EDUROAM credentials... "
     rm /etc/wpa_supplicant/wpa_supplicant.eduroam.conf
-    pkill wpa_supplicant
+    kill -9 $EDUROAM_PID
     iwconfig wlan0 ap 00:00:00:00:00:00
     ifconfig wlan0 0.0.0.0 down
     echo "ok."
@@ -168,34 +189,36 @@ fi
 
 echo -n "Syncing results... "
 if [ ! -z "$IS_INTERNAL" ]; then
-    monroe-rsync-results;
-    rm $BASEDIR/$SCHEDID/container.*
+    monroe-rsync-results
+    rm $_EXPPATH/container.*
 else
-    cat /tmp/cleanup.log > $BASEDIR/$SCHEDID/cleanup.log
+    cat $_tmpfile_ > $_EXPPATH/cleanup.log
     echo "(end of public log)"
-    monroe-user-experiments;  #rsync all remaining files
+    monroe-user-experiments  #rsync all remaining files
 fi
 echo "ok."
 
 echo -n "Cleaning files... "
 # any other file should be rsynced by now
-
-umount $BASEDIR/$SCHEDID
-rmdir  $BASEDIR/$SCHEDID
+#TODO: Report with exist status if cleanup failed
+umount $_EXPPATH
+rmdir  $_EXPPATH
 # .traffic and .status should be kept (are deleted by external scripts after read)
-mv     $STATUSDIR/${SCHEDID}.traffic  $STATUSDIR/${SCHEDID}-traffic
-mv     $STATUSDIR/${SCHEDID}.status  $STATUSDIR/${SCHEDID}-status
-rm -rf $BASEDIR/${SCHEDID}.*
-mv     $STATUSDIR/${SCHEDID}-traffic  $STATUSDIR/${SCHEDID}.traffic_
-mv     $STATUSDIR/${SCHEDID}-status  $STATUSDIR/${SCHEDID}.status
-rm -r  $USAGEDIR/monroe-${SCHEDID}
+mv     $_EXPPATH.traffic  $_EXPPATH-traffic
+mv     $_EXPPATH.status  $_EXPPATH-status
+rm -rf $_EXPPATH.*
+mv     $_EXPPATH-traffic  $_EXPPATH.traffic_
+mv     $_EXPPATH-status  $_EXPPATH.status
+rm -rf  $USAGEDIR/${SCHEDID}
 echo "ok."
 
-echo -n "restorting firewall and modem state"
+echo -n "restarting firewall (and modem state)"
 circle restart
-for ip4table in $(modems|jq .[].ip4table); do
-  curl -s -X POST http://localhost:88/modems/${ip4table}/usbreset;
-done
+if [ -x /usr/bin/modems ]; then 
+  for ip4table in $(modems|jq .[].ip4table); do
+    curl -s -X POST http://localhost:88/modems/${ip4table}/usbreset;
+  done
+fi
 echo "ok."
 
 echo "Cleanup finished $(date)."
